yellow block for rfsoc to synchronize pps with sysref framework
preliminary evaluation suggests it is working as expected. However, the
implementation is more of a hack to work within the toolflow. The functionality
should possibly be handeled better. Because, this specifically works because the
rfsoc platform is expected to be included in the design that provides the
infrastructure for synchronization and this yellow block just relys on it being
present to do anything useful.